// src/app/api/user/model/userInfo.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Friend } from './friend';
import { GrantedAuthority } from './grantedAuthority';

export interface UserInfo { 
    id?: number;
    email?: string;
    firstName?: string;
    lastName?: string;
    nickName?: string;
    password?: string;
    friendList?: Array<Friend>;
    enabled?: boolean;
    username?: string;
    accountNonLocked?: boolean;
    authorities?: Array<GrantedAuthority>;
    credentialsNonExpired?: boolean;
    accountNonExpired?: boolean;
}